backend/requirements.nogit.txt
# Core (pin numpy <2 for compatibility)
numpy==1.26.4

omegaconf==2.1.1
antlr4-python3-runtime==4.8

pytorch-lightning==1.9.5
torchmetrics==0.11.4

einops==0.3.0
kornia==0.6.0

imageio==2.9.0
imageio-ffmpeg==0.4.2
opencv-python==4.9.0.80
albumentations==1.3.1

transformers==4.19.2
torch-fidelity==0.3.0

tqdm
scipy
matplotlib
pillow
ftfy
regex

fastapi
uvicorn[standard]
python-multipart

# Repo deps (non-editable)
